requirements.txt
pillow
mss